ROSA: Prefer dnf over urpmi
Commit 97843a9304 ("fixes for rosa 2019.1 platform") added support for using dnf in rosa2019.1+ where urpmi was changed to dnf.
But there is a converter of urpmi and urpme commands to dnf - dnf-URPM https://github.com/rpm-software-management/dnf-URPM
It will be pre-installed in most ROSA 2019.1-based distros.
Code from commit 97843a9304 will use a converter in this case, what does not make sense.
Use dnf directly if it is available.
There are no cases when dnf is available in urpmi-based ditros, it is just not packaged there.

Fixes: https://github.com/AktivCo/2fa-tuner-lib/pull/1